Add files via upload
This code deals with the development of a controller for the iRobot Create robot in the Webots simulation environment, which uses important functions to avoid collision with obstacles. In this project, an intelligent controller for iRobot Create robot is designed and implemented. This controller uses different sensors to detect obstacles and follow virtual walls. The intelligent controller designed in this project is an important achievement in the field of robotics. This controller can help robots to move safely and precisely in complex environments full of obstacles.